feat(menu): add MultiplayerMenuScene and simplify TitleScene
New multiplayer hub scene with VS EN LINEA, UNIRSE, VS LOCAL,
and TORNEO LOCAL buttons. TitleScene reduced from 8 to 5 buttons.
Join overlay logic moved from TitleScene to MultiplayerMenuScene.
