+++ src/main.js
@@ -12,6 +12,7 @@ import { LeaderboardScene } from './scenes/LeaderboardScene.js';
 import { LearningScene } from './scenes/LearningScene.js';
 import { LobbyScene } from './scenes/LobbyScene.js';
 import { LoginScene } from './scenes/LoginScene.js';
+import { MultiplayerMenuScene } from './scenes/MultiplayerMenuScene.js';
 import { MusicScene } from './scenes/MusicScene.js';
 import { PreFightScene } from './scenes/PreFightScene.js';
 import { SelectScene } from './scenes/SelectScene.js';
@@ -53,6 +54,7 @@ const config = {
     BootScene,
     LoginScene,
     TitleScene,
+    MultiplayerMenuScene,
     TournamentSetupScene,
     BracketScene,
     LobbyScene,
